add pagination, fix footer, resolve issues
