+++ project/src/components/ui/footer/footer.tsx
@@ -1,11 +1,13 @@
+import { Link } from 'react-router-dom';
+
 export const Footer = () => (
   <footer className="page-footer">
     <div className="logo">
-      <a href="main.html" className="logo__link logo__link--light">
+      <Link to="/" className="logo__link logo__link--light">
         <span className="logo__letter logo__letter--1">W</span>
         <span className="logo__letter logo__letter--2">T</span>
         <span className="logo__letter logo__letter--3">W</span>
-      </a>
+      </Link>
     </div>
 
     <div className="copyright">
