+++ project/src/constants/film.ts
@@ -3,3 +3,5 @@ export const TABS = [
   'Details',
   'Reviews',
 ];
+
+export const PAGE_COUNT = 8;
