+++ project/src/pages/main-page/main-page.tsx
@@ -1,28 +1,16 @@
-import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { FilmCard } from '../../components/film-card/film-card';
 import { GenreFilter } from '../../components/genre-filter/genre-filter';
 import { Header } from '../../components/ui/header/header';
-import { ItemList } from '../../components/item-list/item-list';
-import { Loader } from '../../components/ui/loader/loader';
-import { useAppSelector } from '../../hooks/redux-hooks';
 import { TFilm } from '../../types/film';
-import { useFiltredFilms } from '../../hooks/filter-films';
+import { Footer } from '../../components/ui/footer/footer';
+import { ListWithPagination } from '../../components/list-with-pagination/list-with-pagination';
 
 type TMainPageProps = {
   films: TFilm[]
 };
 
 export const MainPage = ({ films }: TMainPageProps) => {
-  const { genre, isLoading } = useAppSelector((state) => state.film);
-  const [ , setFilmId] = useState<number | null>(null);
   const navigate = useNavigate();
-
-  const handleMouseOver = (id: number) => {
-    setFilmId(id);
-  };
-
-  const filtredFilms = useFiltredFilms(genre, films);
 
   return (
     <div>
@@ -32,6 +20,8 @@ export const MainPage = ({ films }: TMainPageProps) => {
         </div>
         <h1 className='visually-hidden'>WTW</h1>
         <Header />
+
+
         <div className='film-card__wrap'>
           <div className='film-card__info'>
             <div className='film-card__poster'>
@@ -63,35 +53,16 @@ export const MainPage = ({ films }: TMainPageProps) => {
             </div>
           </div>
         </div>
-      </section>
 
+
+      </section>
       <div className='page-content'>
         <section className='catalog'>
           <h2 className='catalog__title visually-hidden'>Catalog</h2>
           <GenreFilter films={films} />
-          <div className='catalog__films-list'>
-            {isLoading
-              ? <Loader />
-              : <ItemList items={filtredFilms} renderItem={(item: TFilm) => (<FilmCard film={item} key={item.id} onMouseOver={handleMouseOver} />)} />}
-          </div>
-
-          <div className='catalog__more'>
-            <button className='catalog__button' type='button'>Show more</button>
-          </div>
+          <ListWithPagination films={films}/>
         </section>
-
-        <footer className='page-footer'>
-          <div className='logo'>
-            <a className='logo__link logo__link--light'>
-              <span className='logo__letter logo__letter--1'>W</span>
-              <span className='logo__letter logo__letter--2'>T</span>
-              <span className='logo__letter logo__letter--3'>W</span>
-            </a>
-          </div>
-          <div className='copyright'>
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer />
       </div>
     </div>
   );
