+++ project/src/utils/genres.ts
@@ -1,7 +1,7 @@
 import { TFilm } from '../types/film';
 
 export const generateGenres = (films: TFilm[]): string[]=> {
-  const genres: Set<string> = new Set;
+  const genres: Set<string> = new Set();
 
   genres.add('All genres');
   films.forEach(({genre}) => genres.add(genre));
